increased test coverage and reorganized files
