+++ scripts/__tests__/logger.test.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import logger from '../logger';
+import logger from '../utils/logger';
 
 describe('Logger', () => {
 	const logFilePath = path.resolve('app.log');
