+++ scripts/core/mainStages.ts
@@ -1,0 +1,4 @@
+// Re-export functions from split modules for backward compatibility
+export { loggingIn } from './login.js';
+export { searchForJobs } from './jobSearch.js';
+export { handleMessageApprovalAndApplication } from './application.js';
