+++ scripts/utils/aiUtils.ts
