+++ scripts/utils/debugUtils.ts
