+++ scripts/utils/prompts.ts
